% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles_of_faith.R
\docType{data}
\name{articles_of_faith}
\alias{articles_of_faith}
\title{Articles of Faith}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 13 rows and 19 columns.}
\usage{
articles_of_faith
}
\description{
A small dataset containing the 13 Articles of Faith.
}
\keyword{datasets}
