% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kjv_old_testament.R
\docType{data}
\name{kjv_old_testament}
\alias{kjv_old_testament}
\title{Old Testament (King James Version)}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 23145 rows and 19 columns.}
\source{
\url{http://scriptures.nephi.org/}
}
\usage{
kjv_old_testament
}
\description{
\describe{
  \item{\bold{volume_id}}{ Integer describing unique volume ID}
  \item{\bold{book_id}}{ Integer describing unique book ID}
  \item{\bold{chapter_id}}{ Integer describing unique chapter ID}
  \item{\bold{verse_id}}{ Integer describing unique verse ID}
  \item{\bold{volume_title}}{ Title of volume}
  \item{\bold{book_title}}{ Title of book}
  \item{\bold{volume_long_title}}{ Long Title of Volume}
  \item{\bold{book_long_title}}{ Long Title of Book}
  \item{\bold{volume_subtitle}}{ Volume Subtitle}
  \item{\bold{book_subtitle}}{ Book Subtitle}
  \item{\bold{volume_short_title}}{ Volume Short Title}
  \item{\bold{book_short_title}}{ Book Short Title}
  \item{\bold{volume_lds_url}}{ Follows directory structure to find scriptures on Lds.org}
  \item{\bold{book_lds_url}}{ Follows directory structure to find scriptures on Lds.org}
  \item{\bold{chapter_number}}{ Follows directory structure to find scriptures on Lds.org}
  \item{\bold{verse_number}}{ Follows directory structure to find scriptures on Lds.org}
  \item{\bold{scripture_text}}{ Verse text}
  \item{\bold{verse_title}}{ Human readable verse title (ex. Helaman 3:29)}
  \item{\bold{verse_short_title}}{ Human readable verse title condensed}
}
}
\note{
Find the codebook to this data at \url{http://scriptures.nephi.org/docs.txt}

Find out more about these texts at \url{https://www.lds.org/}
}
\keyword{datasets}
