% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearl_of_great_price.R
\docType{data}
\name{pearl_of_great_price}
\alias{pearl_of_great_price}
\title{Pearl of Great Price}
\format{A tibble with 635 rows and 19 variables
\describe{
  \item{\bold{volume_id}}{ Integer describing unique volume ID}
  \item{\bold{book_id}}{ Integer describing unique book ID}
  \item{\bold{chapter_id}}{ Integer describing unique chapter ID}
  \item{\bold{verse_id}}{ Integer describing unique verse ID}
  \item{\bold{volume_title}}{ Title of volume}
  \item{\bold{book_title}}{ Title of book}
  \item{\bold{volume_long_title}}{ Long Title of Volume}
  \item{\bold{book_long_title}}{ Long Title of Book}
  \item{\bold{volume_subtitle}}{ Volume Subtitle}
  \item{\bold{book_subtitle}}{ Book Subtitle}
  \item{\bold{volume_short_title}}{ Volume Short Title}
  \item{\bold{book_short_title}}{ Book Short Title}
  \item{\bold{volume_lds_url}}{ Follows directory structure to find scriptures on Lds.org}
  \item{\bold{book_lds_url}}{ Follows directory structure to find scriptures on Lds.org}
  \item{\bold{chapter_number}}{ Follows directory structure to find scriptures on Lds.org}
  \item{\bold{verse_number}}{ Follows directory structure to find scriptures on Lds.org}
  \item{\bold{scripture_text}}{ Verse text}
  \item{\bold{verse_title}}{ Human readable verse title (ex. Helaman 3:29)}
  \item{\bold{verse_short_title}}{ Human readable verse title condensed}
}}
\source{
\url{http://scriptures.nephi.org/}
}
\usage{
pearl_of_great_price
}
\description{
A dataset subsetted from \code{scripture_quad} containing all verses in the Pearl of Great Price.
}
\note{
Find the codebook to this data at \url{http://scriptures.nephi.org/docs.txt}

Find out more about these texts at \url{https://www.lds.org/}
}
\keyword{datasets}
